[docs] apply svgr/webpack to storybook
